+++ rt/cpp/db/rt/System.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2007 Digital Bazaar, Inc.  All rights reserved.
+ * Copyright (c) 2007-2008 Digital Bazaar, Inc.  All rights reserved.
  */
 #include "db/rt/System.h"
 
@@ -8,9 +8,9 @@
 
 using namespace db::rt;
 
-unsigned long long System::getCurrentMilliseconds()
+uint64_t System::getCurrentMilliseconds()
 {
-   unsigned long long rval = 0;
+   uint64_t rval = 0;
    
    // get the current time of day
    struct timeval now;
@@ -18,7 +18,7 @@ unsigned long long System::getCurrentMilliseconds()
    
    // get total number of milliseconds
    // 1 millisecond is 1000 microseconds
-   rval = now.tv_sec * 1000LL + (unsigned long)(now.tv_usec / 1000.0);
+   rval = now.tv_sec * 1000ULL + (uint64_t)(now.tv_usec / 1000.0);
    
    return rval;
 }
